Use 600x600 grid size
It's the natural image size and makes is a bit faster to finish since there's less cells to compute.
The current min block size is 30, but it could be 15, not sure how that would look yet.

Also recomputing the dominant color of each album using ColorThief to get albums to match more closely
to the grid cell color.